server/requirements.txt
Flask
torch
torchvision
Pillow
requests